+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added tracking for all known React hooks
 - Added `displayName` field to `fiber-type-def` event
+- Added new section `Hooks` in fiber details
 
 ## 0.7.4 (August 16, 2023)